user_service/requirements.txt
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.3
python-dotenv==1.1.0
python_jose==3.4.0
SQLAlchemy==2.0.40
uvicorn[standard]
psycopg2-binary==2.9.10
